// apps/client/frontend/src/gen/MessageSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AccountId } from "./AccountId";
import type { ConversationId } from "./ConversationId";
import type { MailboxId } from "./MailboxId";
import type { MessageId } from "./MessageId";
import type { Recipient } from "./Recipient";
import type { ThreadId } from "./ThreadId";

/**
 * Twin of [`posthaste_domain_model::MessageSummary`] — the list-row
 * projection.
 */
export type MessageSummary = { id: MessageId, sourceId: AccountId, sourceName: string, sourceThreadId: ThreadId, conversationId: ConversationId, subject: string | null, fromName: string | null, fromEmail: string | null, to: Array<Recipient>, preview: string | null, receivedAt: string, hasAttachment: boolean, isRead: boolean, isFlagged: boolean, mailboxIds: Array<MailboxId>, keywords: Array<string>, version?: number, rfcMessageId?: string, inReplyTo?: string, draftId?: string, };